Spring web ++

+++ spring-web-jpa/src/main/java/no/noroff/accelerate/springjpa/service/StudentServiceImpl.java
@@ -44,6 +44,9 @@ public class StudentServiceImpl implements StudentService {
 
     @Override
     public void deleteById(Integer id) {
+        if (!studentRepository.existsById(id)) {
+            throw new StudentNotFoundException(id);
+        }
         studentRepository.deleteById(id);
     }
